+++ mathics/builtin/arithmetic.py
@@ -7,13 +7,8 @@ Mathematical Functions
 Basic arithmetic functions, including complex number arithmetic.
 """
 
-from mathics.eval.numerify import numerify
-
-# This tells documentation how to sort this module
-sort_order = "mathics.builtin.mathematical-functions"
-
-
 from functools import lru_cache
+from typing import Optional
 
 import mpmath
 import sympy
@@ -49,17 +44,16 @@ from mathics.core.attributes import (
 from mathics.core.convert.expression import to_expression
 from mathics.core.convert.sympy import SympyExpression, from_sympy, sympy_symbol_prefix
 from mathics.core.element import ElementsProperties
+from mathics.core.evaluation import Evaluation
 from mathics.core.expression import Expression
 from mathics.core.list import ListExpression
 from mathics.core.number import dps, min_prec
 from mathics.core.symbols import (
     Atom,
     Symbol,
-    SymbolAbs,
     SymbolFalse,
     SymbolList,
     SymbolPlus,
-    SymbolPower,
     SymbolTimes,
     SymbolTrue,
 )
@@ -72,8 +66,14 @@ from mathics.core.systemsymbols import (
     SymbolTable,
     SymbolUndefined,
 )
-from mathics.eval.arithmetic import eval_mpmath_function
+from mathics.eval.arithmetic import eval_Abs, eval_mpmath_function, eval_Sign
 from mathics.eval.nevaluator import eval_N
+from mathics.eval.numerify import numerify
+
+# This tells documentation how to sort this module
+sort_order = "mathics.builtin.mathematical-functions"
+
+ExpressionComplexInfinity = Expression(SymbolDirectedInfinity)
 
 
 class _MPMathFunction(SympyFunction):
@@ -98,7 +98,7 @@ class _MPMathFunction(SympyFunction):
             return None
         return getattr(mpmath, self.mpmath_name)
 
-    def eval(self, z, evaluation):
+    def eval(self, z, evaluation: Evaluation):
         "%(name)s[z__]"
 
         args = numerify(z, evaluation).get_sequence()
@@ -171,34 +171,35 @@ def create_infix(items, operator, prec, grouping):
 
 class Abs(_MPMathFunction):
     """
-    <url>
-    :Absolute value:
-    https://en.wikipedia.org/wiki/Absolute_value</url> (<url>
-    :SymPy:
-    https://docs.sympy.org/latest/modules/functions/elementary.html#sympy.functions.elementary.complexes.Abs</url>, <url>
-    :WMA: https://reference.wolfram.com/language/ref/Abs</url>)
+        <url>
+        :Absolute value:
+        https://en.wikipedia.org/wiki/Absolute_value</url> (<url>
+        :SymPy:
+        https://docs.sympy.org/latest/modules/functions/
+    elementary.html#sympy.functions.elementary.complexes.Abs</url>, <url>
+        :WMA: https://reference.wolfram.com/language/ref/Abs</url>)
 
-    <dl>
-      <dt>'Abs[$x$]'
-      <dd>returns the absolute value of $x$.
-    </dl>
+        <dl>
+          <dt>'Abs[$x$]'
+          <dd>returns the absolute value of $x$.
+        </dl>
 
-    >> Abs[-3]
-     = 3
+        >> Abs[-3]
+         = 3
 
-    >> Plot[Abs[x], {x, -4, 4}]
-     = -Graphics-
+        >> Plot[Abs[x], {x, -4, 4}]
+         = -Graphics-
 
-    'Abs' returns the magnitude of complex numbers:
-    >> Abs[3 + I]
-     = Sqrt[10]
-    >> Abs[3.0 + I]
-     = 3.16228
+        'Abs' returns the magnitude of complex numbers:
+        >> Abs[3 + I]
+         = Sqrt[10]
+        >> Abs[3.0 + I]
+         = 3.16228
 
-    All of the below evaluate to Infinity:
+        All of the below evaluate to Infinity:
 
-    >> Abs[Infinity] == Abs[I Infinity] == Abs[ComplexInfinity]
-     = True
+        >> Abs[Infinity] == Abs[I Infinity] == Abs[ComplexInfinity]
+         = True
     """
 
     mpmath_name = "fabs"  # mpmath actually uses python abs(x) / x.__abs__()
@@ -207,6 +208,13 @@ class Abs(_MPMathFunction):
     }
     summary_text = "absolute value of a number"
     sympy_name = "Abs"
+
+    def eval(self, x, evaluation: Evaluation):
+        "Abs[x_]"
+        result = eval_Abs(x)
+        if result is not None:
+            return result
+        return super(Abs, self).eval(x, evaluation)
 
 
 class Arg(_MPMathFunction):
@@ -264,7 +272,7 @@ class Arg(_MPMathFunction):
     sympy_name = "arg"
 
     def eval(self, z, evaluation, options={}):
-        "%(name)s[z_, OptionsPattern[%(name)s]]"
+        "Arg[z_, OptionsPattern[Arg]]"
         if Expression(SymbolPossibleZeroQ, z).evaluate(evaluation) is SymbolTrue:
             return Integer0
         preference = self.get_option(options, "Method", evaluation).get_string_value()
@@ -302,7 +310,7 @@ class Assuming(Builtin):
     summary_text = "set assumptions during the evaluation"
     attributes = A_HOLD_REST | A_PROTECTED
 
-    def eval_assuming(self, assumptions, expr, evaluation):
+    def eval_assuming(self, assumptions, expr, evaluation: Evaluation):
         "Assuming[assumptions_, expr_]"
         assumptions = assumptions.evaluate(evaluation)
         if assumptions is SymbolTrue:
@@ -358,11 +366,11 @@ class Boole(Builtin):
      = Boole[a == 7]
     """
 
-    summary_text = "translate 'True' to 1, and 'False' to 0"
     attributes = A_LISTABLE | A_PROTECTED
+    summary_text = "translate 'True' to 1, and 'False' to 0"
 
-    def eval(self, expr, evaluation):
-        "%(name)s[expr_]"
+    def eval(self, expr, evaluation: Evaluation):
+        "Boole[expr_]"
         if expr is SymbolTrue:
             return Integer1
         elif expr is SymbolFalse:
@@ -432,8 +440,8 @@ class Complex_(Builtin):
     summary_text = "head for complex numbers"
     name = "Complex"
 
-    def eval(self, r, i, evaluation):
-        "%(name)s[r_?NumberQ, i_?NumberQ]"
+    def eval(self, r, i, evaluation: Evaluation):
+        "Complex[r_?NumberQ, i_?NumberQ]"
 
         if isinstance(r, Complex) or isinstance(i, Complex):
             sym_form = r.to_sympy() + sympy.I * i.to_sympy()
@@ -444,11 +452,13 @@ class Complex_(Builtin):
 
 class ConditionalExpression(Builtin):
     """
-    <url>:WMA link:https://reference.wolfram.com/language/ref/ConditionalExpression.html</url>
+    <url>:WMA link:https://reference.wolfram.com/
+language/ref/ConditionalExpression.html</url>
 
     <dl>
       <dt>'ConditionalExpression[$expr$, $cond$]'
-      <dd>returns $expr$ if $cond$ evaluates to $True$, $Undefined$ if $cond$ evaluates to $False$.
+      <dd>returns $expr$ if $cond$ evaluates to $True$, $Undefined$ if $cond$ \
+          evaluates to $False$.
     </dl>
 
     >> ConditionalExpression[x^2, True]
@@ -485,7 +495,7 @@ class ConditionalExpression(Builtin):
         "expr1_ ^ ConditionalExpression[expr2_, cond_]": "ConditionalExpression[expr1^expr2, cond]",
     }
 
-    def eval_generic(self, expr, cond, evaluation):
+    def eval_generic(self, expr, cond, evaluation: Evaluation):
         "ConditionalExpression[expr_, cond_]"
         # What we need here is a way to evaluate
         # cond as a predicate, using assumptions.
@@ -590,8 +600,7 @@ class DirectedInfinity(SympyFunction):
      = Indeterminate
 
     >> DirectedInfinity[0]
-     : Indeterminate expression 0 Infinity encountered.
-     = Indeterminate
+     = ComplexInfinity
 
     #> DirectedInfinity[1+I]+DirectedInfinity[2+I]
      = (2 / 5 + I / 5) Sqrt[5] Infinity + (1 / 2 + I / 2) Sqrt[2] Infinity
@@ -602,15 +611,12 @@ class DirectedInfinity(SympyFunction):
 
     summary_text = "infinite quantity with a defined direction in the complex plane"
     rules = {
-        "DirectedInfinity[Indeterminate]": "Indeterminate",
         "DirectedInfinity[args___] ^ -1": "0",
-        "0 * DirectedInfinity[args___]": "Message[Infinity::indet, Unevaluated[0 DirectedInfinity[args]]]; Indeterminate",
-        # "DirectedInfinity[a_?NumericQ] /; N[Abs[a]] != 1": "DirectedInfinity[a / Abs[a]]",
-        # "DirectedInfinity[a_] * DirectedInfinity[b_]": "DirectedInfinity[a*b]",
-        # "DirectedInfinity[] * DirectedInfinity[args___]": "DirectedInfinity[]",
-        # Rules already implemented in Times.eval
-        #        "z_?NumberQ * DirectedInfinity[]": "DirectedInfinity[]",
-        #        "z_?NumberQ * DirectedInfinity[a_]": "DirectedInfinity[z * a]",
+        # Special arguments:
+        "DirectedInfinity[DirectedInfinity[args___]]": "DirectedInfinity[args]",
+        "DirectedInfinity[Indeterminate]": "Indeterminate",
+        "DirectedInfinity[Alternatives[0, 0.]]": "DirectedInfinity[]",
+        # Plus
         "DirectedInfinity[a_] + DirectedInfinity[b_] /; b == -a": (
             "Message[Infinity::indet,"
             "  Unevaluated[DirectedInfinity[a] + DirectedInfinity[b]]];"
@@ -622,17 +628,15 @@ class DirectedInfinity(SympyFunction):
             "Indeterminate"
         ),
         "DirectedInfinity[args___] + _?NumberQ": "DirectedInfinity[args]",
-        "DirectedInfinity[0]": (
+        # Times. See if can be reinstalled in eval_Times
+        "Alternatives[0, 0.] DirectedInfinity[z___]": (
             "Message[Infinity::indet,"
-            "  Unevaluated[DirectedInfinity[0]]];"
+            "  Unevaluated[0 DirectedInfinity[z]]];"
             "Indeterminate"
         ),
-        "DirectedInfinity[0.]": (
-            "Message[Infinity::indet,"
-            "  Unevaluated[DirectedInfinity[0.]]];"
-            "Indeterminate"
-        ),
-        "DirectedInfinity[DirectedInfinity[x___]]": "DirectedInfinity[x]",
+        "a_?NumericQ * DirectedInfinity[b_]": "DirectedInfinity[a * b]",
+        "a_ DirectedInfinity[]": "DirectedInfinity[]",
+        "DirectedInfinity[a_] * DirectedInfinity[b_]": "DirectedInfinity[a * b]",
     }
 
     formats = {
@@ -643,18 +647,41 @@ class DirectedInfinity(SympyFunction):
         "DirectedInfinity[z_?NumericQ]": "HoldForm[z Infinity]",
     }
 
-    def eval(self, z, evaluation):
-        """DirectedInfinity[z_]"""
-        if z in (Integer1, IntegerM1):
+    def eval_directed_infinity(self, direction, evaluation: Evaluation):
+        """DirectedInfinity[direction_]"""
+        if direction in (Integer1, IntegerM1):
             return None
-        if isinstance(z, Number) or isinstance(eval_N(z, evaluation), Number):
-            direction = (z / Expression(SymbolAbs, z)).evaluate(evaluation)
-            return Expression(
-                SymbolDirectedInfinity,
-                direction,
-                elements_properties=ElementsProperties(True, True, True),
-            )
-        return None
+        if direction.is_zero:
+            return ExpressionComplexInfinity
+
+        normalized_direction = eval_Sign(direction)
+        # TODO: improve eval_Sign, to avoid the need of the
+        # following block:
+        #   ############################################
+        if normalized_direction is None:
+            ndir = eval_N(direction, evaluation)
+            if isinstance(ndir, (Integer, Rational, Real)):
+                if abs(ndir.value) == 1.0:
+                    normalized_direction = direction
+                else:
+                    normalized_direction = direction / Abs(direction)
+            elif isinstance(ndir, Complex):
+                re, im = ndir.value
+                if abs(re.value**2 + im.value**2 - 1.0) < 1.0e-9:
+                    normalized_direction = direction
+                else:
+                    normalized_direction = direction / Abs(direction)
+            else:
+                return None
+        #  ##############################################
+
+        if normalized_direction is None:
+            return None
+        return Expression(
+            SymbolDirectedInfinity,
+            normalized_direction.evaluate(evaluation),
+            elements_properties=ElementsProperties(True, False, False),
+        )
 
     def to_sympy(self, expr, **kwargs):
         if len(expr.elements) == 1:
@@ -689,7 +716,7 @@ class I_(Predefined):
     summary_text = "imaginary unit"
     python_equivalent = 1j
 
-    def evaluate(self, evaluation):
+    def evaluate(self, evaluation: Evaluation):
         return Complex(Integer0, Integer1)
 
 
@@ -717,17 +744,17 @@ class Im(SympyFunction):
     summary_text = "imaginary part"
     attributes = A_LISTABLE | A_NUMERIC_FUNCTION | A_PROTECTED
 
-    def eval_complex(self, number, evaluation):
+    def eval_complex(self, number, evaluation: Evaluation):
         "Im[number_Complex]"
+        if isinstance(number, Complex):
+            return number.imag
 
-        return number.imag
-
-    def eval_number(self, number, evaluation):
+    def eval_number(self, number, evaluation: Evaluation):
         "Im[number_?NumberQ]"
 
         return Integer0
 
-    def eval(self, number, evaluation):
+    def eval(self, number, evaluation: Evaluation):
         "Im[number_]"
 
         return from_sympy(sympy.im(number.to_sympy().expand(complex=True)))
@@ -795,7 +822,7 @@ class Piecewise(SympyFunction):
 
     attributes = A_HOLD_ALL | A_PROTECTED
 
-    def eval(self, items, evaluation):
+    def eval(self, items, evaluation: Evaluation):
         "%(name)s[items__]"
         result = self.to_sympy(
             Expression(SymbolPiecewise, *items.get_sequence()), evaluation=evaluation
@@ -954,8 +981,8 @@ class Rational_(Builtin):
     summary_text = "head for rational numbers"
     name = "Rational"
 
-    def eval(self, n: Integer, m: Integer, evaluation):
-        "%(name)s[n_Integer, m_Integer]"
+    def eval(self, n: Integer, m: Integer, evaluation: Evaluation):
+        "Rational[n_Integer, m_Integer]"
 
         if m.value == 1:
             return n
@@ -988,19 +1015,18 @@ class Re(SympyFunction):
     attributes = A_LISTABLE | A_NUMERIC_FUNCTION | A_PROTECTED
     sympy_name = "re"
 
-    def eval_complex(self, number, evaluation):
+    def eval_complex(self, number, evaluation: Evaluation):
         "Re[number_Complex]"
+        if isinstance(number, Complex):
+            return number.real
 
-        return number.real
-
-    def eval_number(self, number, evaluation):
+    def eval_number(self, number, evaluation: Evaluation):
         "Re[number_?NumberQ]"
 
         return number
 
-    def eval(self, number, evaluation):
+    def eval(self, number, evaluation: Evaluation):
         "Re[number_]"
-
         return from_sympy(sympy.re(number.to_sympy().expand(complex=True)))
 
 
@@ -1148,22 +1174,24 @@ class Sign(SympyFunction):
         "argx": "Sign called with `1` arguments; 1 argument is expected.",
     }
 
-    def eval(self, x, evaluation):
+    rules = {
+        "Sign[Power[a_, b_]]": "Power[Sign[a], b]",
+    }
+
+    def eval(self, x, evaluation: Evaluation):
         "%(name)s[x_]"
-        # Sympy and mpmath do not give the desired form of complex number
-        if isinstance(x, Complex):
-            return Expression(
-                SymbolTimes,
-                x,
-                Expression(SymbolPower, Expression(SymbolAbs, x), IntegerM1),
-            )
+        result = eval_Sign(x)
+        if result is not None:
+            return result
+        # return None
 
         sympy_x = x.to_sympy()
         if sympy_x is None:
             return None
-        return super().eval(x, evaluation)
+        # Unhandled cases. Use sympy
+        return super(Sign, self).eval(x, evaluation)
 
-    def eval_error(self, x, seqs, evaluation):
+    def eval_error(self, x, seqs, evaluation: Evaluation):
         "Sign[x_, seqs__]"
         evaluation.message("Sign", "argx", Integer(len(seqs.get_sequence()) + 1))
 
@@ -1183,7 +1211,8 @@ class Sum(IterationFunction, SympyFunction):
       <dd>$i$ ranges from $imin$ to $imax$ in steps of $di$.
 
       <dt>'Sum[$expr$, {$i$, $imin$, $imax$}, {$j$, $jmin$, $jmax$}, ...]'
-      <dd>evaluates $expr$ as a multiple sum, with {$i$, ...}, {$j$, ...}, ... being in outermost-to-innermost order.
+      <dd>evaluates $expr$ as a multiple sum, with {$i$, ...}, {$j$, ...}, ... being \
+          in outermost-to-innermost order.
     </dl>
 
 
@@ -1269,7 +1298,7 @@ class Sum(IterationFunction, SympyFunction):
     def get_result(self, items):
         return Expression(SymbolPlus, *items)
 
-    def to_sympy(self, expr, **kwargs) -> SympyExpression:
+    def to_sympy(self, expr, **kwargs) -> Optional[SympyExpression]:
         """
         Perform summation via sympy.summation
         """